refactor:PlayerContext
