+++ podcastr/src/contexts/PlayerContext.tsx
@@ -1,0 +1,61 @@
+import { createContext, useState, ReactNode } from 'react';
+
+type EpisodeType = {
+  title: string;
+  members: string;
+  thumbnail: string;
+  duration: number;
+  url: string;
+};
+
+type PlayerContextData = {
+  episodeList: EpisodeType[];
+  currentEpisodeIndex: number;
+  isPlaying: boolean;
+  play: (episode: EpisodeType) => void;
+  togglePlay: () => void;
+  setPlayingState: (state: boolean) => void;
+};
+
+export const PlayerContext = createContext({} as PlayerContextData);
+
+type PlayerContextProviderProps = {
+  children: ReactNode;
+};
+
+export function PlayerContextProvider({
+  children,
+}: PlayerContextProviderProps) {
+  const [episodeList, setEpisodeList] = useState([]);
+  const [currentEpisodeIndex, setCurrentEpisodeIndex] = useState(0);
+  const [isPlaying, setIsPlaying] = useState(false);
+
+  function play(episode: EpisodeType) {
+    setEpisodeList([episode]);
+    setCurrentEpisodeIndex(0);
+    setIsPlaying(true);
+  }
+
+  function togglePlay() {
+    setIsPlaying(!isPlaying);
+  }
+
+  function setPlayingState(state: boolean) {
+    setIsPlaying(state);
+  }
+
+  return (
+    <PlayerContext.Provider
+      value={{
+        episodeList,
+        currentEpisodeIndex,
+        play,
+        isPlaying,
+        togglePlay,
+        setPlayingState,
+      }}
+    >
+      {children}
+    </PlayerContext.Provider>
+  );
+}
